Day 9: simple parser
I spent more time finding a file input function I was happy with to suck in the data stream this time around, having forgotten I'd used line by line file I/O (do people still spell it that way?) on Day 8. Minor tidying of code for days 7 and 8 for readability as well.

The parsing was straightforward based on strict rules. The tricky bit was handling cancelled characters, which I do in the code by handling the ! state separately from processing the stream.