Day 20: particles in 3d space
Starting with a set of particles, each with an initial position, initial velocity, and constant acceleration:
 - part 1: find the particle that will remain closest to the origin in the long term
 - part 2: remove all particles from every collision and return the number of particles that remain in the long term
Yes, I ran for 10 times more iterations than I needed to, to ensure my answers remained stable. Wasn't in a mood to write "same for N generations, assume it's the same forever" into the code when I saw the answer for the first part seemed stable but then changed after several iterations.