Day 25: Turing machine
Run the give state machine for a large number of turns and produce the sum of all '1' bits.
I took the belt and suspenders approach of verifying the checksum in case I typo'd the input. Turns out in my first run I'd typed a different part of the input incorrectly.